# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmedFacts": {
        "additionalProperties": false,
        "properties": {
          "age": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "family_income_inr": {
            "anyOf": [
              {
                "maximum": 1000000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Income Inr"
          },
          "gender": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "has_disability": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Disability"
          },
          "is_student": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Student"
          },
          "land_area_acres": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Land Area Acres"
          },
          "land_registration": {
            "anyOf": [
              {
                "enum": [
                  "yes",
                  "no",
                  "not_sure"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Land Registration"
          },
          "occupation": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "social_category": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Social Category"
          },
          "state_code": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          }
        },
        "title": "ConfirmedFacts",
        "type": "object"
      },
      "ConfirmedMatchRequest": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "$ref": "#/components/schemas/ConfirmedFacts"
          },
          "limit": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "facts"
        ],
        "title": "ConfirmedMatchRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchesResponse": {
        "additionalProperties": false,
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SchemeMatchResult"
            },
            "title": "Results",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "results"
        ],
        "title": "MatchesResponse",
        "type": "object"
      },
      "NextQuestionRequest": {
        "additionalProperties": false,
        "properties": {
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "run_id"
        ],
        "title": "NextQuestionRequest",
        "type": "object"
      },
      "NextQuestionResponse": {
        "additionalProperties": false,
        "properties": {
          "answer_type": {
            "anyOf": [
              {
                "enum": [
                  "single_choice",
                  "number",
                  "text"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Type"
          },
          "field": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "title": "Options",
            "type": "array"
          },
          "question": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "NextQuestionResponse",
        "type": "object"
      },
...
      "ProfileExtractRequest": {
        "additionalProperties": false,
        "properties": {
          "locale": {
            "default": "en-IN",
            "pattern": "^[a-z]{2}(?:-[A-Z]{2})?$",
            "title": "Locale",
            "type": "string"
          },
          "text": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ProfileExtractRequest",
        "type": "object"
      },
      "ProfileExtractResponse": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "$ref": "#/components/schemas/ProfileFacts"
          },
          "needs_review": {
            "title": "Needs Review",
            "type": "boolean"
          },
          "unknown_fields": {
            "items": {
              "type": "string"
            },
            "title": "Unknown Fields",
            "type": "array"
          }
        },
        "required": [
          "facts",
          "unknown_fields",
          "needs_review"
        ],
        "title": "ProfileExtractResponse",
        "type": "object"
      },
      "ProfileFacts": {
        "additionalProperties": false,
        "description": "Normalized facts; omitted citizen information remains null.",
        "properties": {
          "age": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "family_income_inr": {
            "anyOf": [
              {
                "maximum": 1000000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Income Inr"
          },
          "gender": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "has_disability": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Disability"
          },
          "is_student": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Student"
          },
          "land_area_acres": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Land Area Acres"
          },
          "land_registration": {
            "anyOf": [
              {
                "enum": [
                  "yes",
                  "no",
                  "not_sure"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Land Registration"
          },
          "occupation": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupation"
          },
          "social_category": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Social Category"
          },
          "state_code": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          }
        },
        "title": "ProfileFacts",
        "type": "object"
      },
...
      "RuleOutcome": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "required_field": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Field"
          },
          "result": {
            "$ref": "#/components/schemas/RuleResult"
          },
          "rule_key": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Rule Key",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "rule_key",
          "result",
          "source_id",
          "reason"
        ],
        "title": "RuleOutcome",
        "type": "object"
      },
      "RuleResult": {
        "enum": [
          "pass",
          "fail",
          "unknown",
          "manual_review"
        ],
        "title": "RuleResult",
        "type": "string"
      },
...
      "SchemeMatchResult": {
        "additionalProperties": false,
        "properties": {
          "failed_rules": {
            "items": {
              "$ref": "#/components/schemas/RuleOutcome"
            },
            "title": "Failed Rules",
            "type": "array"
          },
          "last_verified_at": {
            "format": "date-time",
            "title": "Last Verified At",
            "type": "string"
          },
          "manual_review_rules": {
            "items": {
              "$ref": "#/components/schemas/RuleOutcome"
            },
            "title": "Manual Review Rules",
            "type": "array"
          },
          "matched_rules": {
            "items": {
              "$ref": "#/components/schemas/RuleOutcome"
            },
            "title": "Matched Rules",
            "type": "array"
          },
          "official_source_urls": {
            "items": {
              "format": "uri",
              "minLength": 1,
              "type": "string"
            },
            "minItems": 1,
            "title": "Official Source Urls",
            "type": "array"
          },
          "relevance_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance Score",
            "type": "number"
          },
          "scheme_id": {
            "format": "uuid",
            "title": "Scheme Id",
            "type": "string"
          },
          "scheme_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Scheme Name",
            "type": "string"
          },
          "scheme_version_id": {
            "format": "uuid",
            "title": "Scheme Version Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Verdict"
          },
          "unknown_rules": {
            "items": {
              "$ref": "#/components/schemas/RuleOutcome"
            },
            "title": "Unknown Rules",
            "type": "array"
          }
        },
        "required": [
          "scheme_id",
          "scheme_name",
          "scheme_version_id",
          "status",
          "relevance_score",
          "matched_rules",
          "failed_rules",
          "unknown_rules",
          "last_verified_at",
          "official_source_urls"
        ],
        "title": "SchemeMatchResult",
        "type": "object"
      },
...
      "StagingAskRequest": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "default": 6,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "question": {
            "maxLength": 500,
            "minLength": 2,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "StagingAskRequest",
        "type": "object"
      },
      "StagingAskResponse": {
        "additionalProperties": false,
        "description": "Generated from unverified drafts for curator triage; never citizen guidance.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "cited_slugs": {
            "items": {
              "type": "string"
            },
            "title": "Cited Slugs",
            "type": "array"
          },
          "publication_allowed": {
            "const": false,
            "default": false,
            "title": "Publication Allowed",
            "type": "boolean"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/StagingSearchResult"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "cited_slugs",
          "sources"
        ],
        "title": "StagingAskResponse",
        "type": "object"
      },
      "StagingSearchResponse": {
        "additionalProperties": false,
        "description": "Unverified curation leads; publication_allowed is always false.",
        "properties": {
          "publication_allowed": {
            "const": false,
            "default": false,
            "title": "Publication Allowed",
            "type": "boolean"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/StagingSearchResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "StagingSearchResponse",
        "type": "object"
      },
      "StagingSearchResult": {
        "additionalProperties": false,
        "properties": {
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "title": "Missing Fields",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "record": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Record",
            "type": "object"
          },
          "review_status": {
            "const": "draft",
            "default": "draft",
            "title": "Review Status",
            "type": "string"
          },
          "similarity": {
            "title": "Similarity",
            "type": "number"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "similarity",
          "missing_fields",
          "record"
        ],
        "title": "StagingSearchResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verdict": {
        "enum": [
          "all_checked_conditions_met",
          "needs_information",
          "not_eligible",
          "manual_review"
        ],
        "title": "Verdict",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Yojana Saathi API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/staging-schemes/ask": {
      "post": {
        "description": "Answer a curator question from retrieved unverified records, with checked citations.",
        "operationId": "ask_staging_schemes_api_v1_admin_staging_schemes_ask_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StagingAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StagingAskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask Staging Schemes",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/staging-schemes/search": {
      "get": {
        "description": "Find unverified notebook records worth curating. Never used for citizen matching.",
        "operationId": "search_staging_schemes_api_v1_admin_staging_schemes_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StagingSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Staging Schemes",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/matches": {
      "post": {
        "operationId": "matches_api_v1_matches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmedMatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Matches",
        "tags": [
          "matching"
        ]
      }
    },
...
    "/api/v1/profiles/extract": {
      "post": {
        "operationId": "extract_api_v1_profiles_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileExtractResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract",
        "tags": [
          "profiles"
        ]
      }
    },
...
    "/api/v1/questions/next": {
      "post": {
        "operationId": "questions_api_v1_questions_next_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextQuestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Questions",
        "tags": [
          "questions"
        ]
      }
    },
...
  }
}